% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{add_credential_function}
\alias{add_credential_function}
\title{Add a new credential fetching function.}
\usage{
add_credential_function(f)
}
\arguments{
\item{f}{A function with the right signature. See \code{\link[=is_credential_function]{is_credential_function()}}.}
}
\description{
Note that this implicitly adds \code{f} to the \emph{end} of the list.
}
\seealso{
Other registration: \code{\link{all_credential_functions}},
  \code{\link{set_credential_functions}}
}
