% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{all_credential_functions}
\alias{all_credential_functions}
\title{Get the list of all credential functions.}
\usage{
all_credential_functions()
}
\value{
A list of credential functions.
}
\description{
Get the list of all credential functions.
}
\seealso{
Other registration: \code{\link{add_credential_function}},
  \code{\link{clear_credential_functions}},
  \code{\link{set_credential_functions}}
}
