% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-state.R
\name{auth_activate}
\alias{auth_activate}
\alias{auth_deactivate}
\alias{auth_status}
\title{Activate / de-activate auth}
\usage{
auth_activate()

auth_deactivate(clear = FALSE)

auth_status()
}
\value{
the auth state, invisible
}
\description{
Activate / de-activate auth
}
