% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{basic_welcome_ui}
\alias{basic_welcome_ui}
\title{Helper function for creating a basic welcome screen}
\usage{
basic_welcome_ui(...)
}
\arguments{
\item{...}{\emph{Unnamed} arguments should be Shiny UI objects (i.e. HTML tags), and
will become the page's main contents; they will immediately be followed by
a Google signin button.

\emph{Named} arguments become attributes on the innermost
\verb{<div>} element that wraps both the given content, and the sign-in button.}
}
\description{
Call this function with a bit of content (say, the title of your app and a
couple of sentences describing why login is required) and a \code{welcome_ui}
function will be returned, suitable for passing to \code{\link[=require_oauth]{require_oauth()}}. (See
the Details section of \code{\link[=require_oauth]{require_oauth()}} to see an example of
\code{basic_welcome_ui}.)
}
