% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{bulletize}
\alias{bulletize}
\title{Abbreviate a bullet list neatly}
\usage{
bulletize(x, bullet = "*", n_show = 5, n_fudge = 2)
}
\description{
For internal use in gargle, googledrive, and googlesheets4 (for now).
}
\keyword{internal}
