% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{clear_credential_functions}
\alias{clear_credential_functions}
\title{Clear the list of credential functions.}
\usage{
clear_credential_functions()
}
\description{
Clear the list of credential functions.
}
\seealso{
Other registration: \code{\link{add_credential_function}},
  \code{\link{all_credential_functions}},
  \code{\link{set_credential_functions}}
}
