% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-cookies.R
\name{cookie_options}
\alias{cookie_options}
\title{HTTP cookie options}
\usage{
cookie_options(
  max_age = NULL,
  domain = NULL,
  path = NULL,
  secure = NULL,
  http_only = TRUE,
  same_site = NULL,
  expires = NULL
)
}
\arguments{
\item{max_age}{Either \code{NULL} or a number indicating how many seconds (after a
cookie is set) until the cookie expires. If both \code{expires} and \code{max_age}
are \code{NULL}, then the cookie will be removed when the browser shuts down.}

\item{domain, path, secure, same_site}{Standard HTTP cookie options; see
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie}{MDN}
for details.}

\item{http_only}{Either \code{NULL} or \code{TRUE}, which indicates that the cookie
should not be readable by JavaScript code in the client, only by the server
(the R process running Shiny). You should use \code{TRUE} unless you know you
have a specific reason not to.}

\item{expires}{A \link[base:DateTimeClasses]{POSIXt} (\code{POSIXlt} or \code{POSIXct}) object
specifying the time at which the cookie expires. (You'll generally want to
use \code{max_age} instead; and if both \code{max_age} and \code{expires} are both
specified, \code{max_age} takes precedence.)}
}
\description{
Creates a cookie options object that can be passed to \code{\link[=require_oauth]{require_oauth()}},
to use when writing HTTP cookies for persisting auth credentials.
}
