% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cred_funs_add}
\alias{cred_funs_add}
\title{Add a new credential fetching function.}
\usage{
cred_funs_add(...)
}
\arguments{
\item{...}{One or more functions with the right signature. See
\code{\link[=is_cred_fun]{is_cred_fun()}}.}
}
\description{
Function(s) are added to the \emph{front} of the list.
}
\examples{
creds_one <- function(scopes, ...) {}
cred_funs_add(creds_one)
cred_funs_add(one = creds_one)
cred_funs_add(one = creds_one, two = creds_one)
cred_funs_add(one = creds_one, creds_one)
}
\seealso{
Other registration: \code{\link{cred_funs_clear}},
  \code{\link{cred_funs_list}}, \code{\link{cred_funs_set}}
}
