% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-function-registry.R
\name{cred_funs_clear}
\alias{cred_funs_clear}
\title{Clear the credential function registry}
\usage{
cred_funs_clear()
}
\description{
Clear the credential function registry
}
\seealso{
Other registry management: \code{\link{cred_funs_add}},
  \code{\link{cred_funs_list}}, \code{\link{cred_funs_set}}
}
\concept{registry management}
