% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cred_funs_list}
\alias{cred_funs_list}
\title{List the registered credential functions}
\usage{
cred_funs_list()
}
\value{
A list of credential functions.
}
\description{
List the registered credential functions
}
\seealso{
Other registration: \code{\link{cred_funs_add}},
  \code{\link{cred_funs_clear}},
  \code{\link{cred_funs_set}}
}
\concept{registration}
