% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cred_funs_set}
\alias{cred_funs_set}
\title{Register a list of credential fetching functions}
\usage{
cred_funs_set(ls)
}
\arguments{
\item{ls}{A list of credential functions.}
}
\description{
Register a list of credential fetching functions
}
\seealso{
Other registration: \code{\link{cred_funs_add}},
  \code{\link{cred_funs_clear}},
  \code{\link{cred_funs_list}}
}
\concept{registration}
