% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-function-registry.R
\name{cred_funs_set_default}
\alias{cred_funs_set_default}
\title{Set the default credential functions}
\usage{
cred_funs_set_default()
}
\description{
Set the default credential functions
}
