% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_connect.R
\name{credentials_connect}
\alias{credentials_connect}
\title{Get a token on Posit Connect}
\usage{
credentials_connect(scopes = NULL, ...)
}
\arguments{
\item{scopes}{A character vector of scopes to request. Pick from those listed
at \url{https://developers.google.com/identity/protocols/oauth2/scopes}.

For certain token flows, the
\code{"https://www.googleapis.com/auth/userinfo.email"} scope is unconditionally
included. This grants permission to retrieve the email address associated
with a token; gargle uses this to index cached OAuth tokens. This grants no
permission to view or send email and is generally considered a low-value
scope.}

\item{...}{Additional arguments passed to all credential functions.}
}
\value{
A \code{\link[httr:Token-class]{httr::Token2.0()}} or \code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Shiny apps running on Posit Connect \href{https://docs.posit.co/connect/user/oauth-integrations/}{can retrieve Google credentials for each individual viewer}.

Requires the \pkg{connectcreds} package.
}
\examples{
credentials_connect()
}
\seealso{
Other credential functions: 
\code{\link{credentials_app_default}()},
\code{\link{credentials_byo_oauth2}()},
\code{\link{credentials_external_account}()},
\code{\link{credentials_gce}()},
\code{\link{credentials_service_account}()},
\code{\link{credentials_user_oauth2}()},
\code{\link{token_fetch}()}
}
\concept{credential functions}
