% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-credentials.R
\name{credentials_travis}
\alias{credentials_travis}
\title{Create credentials from a service account token stored by travis.}
\usage{
credentials_travis(scopes, path = "", ...)
}
\arguments{
\item{scopes}{A character vector of scopes to request. Pick from those listed
in \href{https://developers.google.com/identity/protocols/googlescopes}{OAuth 2.0 Scopes for Google APIs}. For
certain token flows, the \code{"https://www.googleapis.com/auth/userinfo.email"}
scope may be added, if not already present. This grants permission to view
the email address associated with a token; gargle uses this to index cached
OAuth tokens. This grants no permission to view or send email. It is
considered a low value scope and does not appear on the consent screen.}

\item{path}{Path to the decrypted travis service account.}

\item{...}{Additional arguments passed to all credentials functions.}
}
\description{
Create credentials from a service account token stored by travis.
}
\seealso{
Other credential functions: \code{\link{credentials_app_default}},
  \code{\link{credentials_gce}},
  \code{\link{credentials_service_account}},
  \code{\link{credentials_user_oauth2}},
  \code{\link{token_fetch}}
}
\concept{credential functions}
