% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{credfuns_add}
\alias{credfuns_add}
\title{Add a new credential fetching function.}
\usage{
credfuns_add(...)
}
\arguments{
\item{...}{One or more functions with the right signature. See
\code{\link[=is_credfun]{is_credfun()}}.}
}
\description{
Function(s) are added to the \emph{front} of the list.
}
\examples{
creds_one <- function(scopes, ...) {}
credfuns_add(creds_one)
credfuns_add(one = creds_one)
credfuns_add(one = creds_one, two = creds_one)
credfuns_add(one = creds_one, creds_one)
}
\seealso{
Other registration: \code{\link{credfuns_clear}},
  \code{\link{credfuns_list}}, \code{\link{credfuns_set}}
}
