% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{credfuns_clear}
\alias{credfuns_clear}
\title{Clear the list of credential functions.}
\usage{
credfuns_clear()
}
\description{
Clear the list of credential functions.
}
\seealso{
Other registration: \code{\link{credfuns_add}},
  \code{\link{credfuns_list}}, \code{\link{credfuns_set}}
}
