% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{credfuns_list}
\alias{credfuns_list}
\title{Get the list of all credential functions.}
\usage{
credfuns_list()
}
\value{
A list of credential functions.
}
\description{
Get the list of all credential functions.
}
\seealso{
Other registration: \code{\link{credfuns_add}},
  \code{\link{credfuns_clear}}, \code{\link{credfuns_set}}
}
