% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{credfuns_set}
\alias{credfuns_set}
\title{Set the list of all credential functions.}
\usage{
credfuns_set(ls)
}
\arguments{
\item{ls}{A list of credential functions.}
}
\description{
Set the list of all credential functions.
}
\seealso{
Other registration: \code{\link{credfuns_add}},
  \code{\link{credfuns_clear}}, \code{\link{credfuns_list}}
}
