% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expose}
\alias{expose}
\alias{is_expose}
\title{An expose object}
\usage{
expose()

is_expose(x)
}
\arguments{
\item{x}{An object that might be an \code{expose} object.}
}
\description{
\code{expose()} returns a sentinel object, similar in spirit to \code{NULL}, that tells
the calling function to return its internal data source. Client packages such
as googledrive and googlesheets4 store information internally about the
target API, such as Drive file MIME types, and then expose this via helper
functions, like \code{googledrive::drive_mime_type()} and
\code{googlesheets4::sheets_endpoints()}. These internal objects are used to
provide nice defaults, check input validity, or lookup something cryptic,
like MIME type, based on something friendlier, like a file extension. Pass
\code{expose()} to such a function to see the internal object, in its full glory.
This is inspired by the \code{waiver()} object in ggplot2.
}
\examples{
\dontrun{
googledrive::drive_mime_type(expose())
googledrive::drive_fields(expose())

is_expose(expose())
is_expose("nope")
}

}
