% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-package.R, R/ui.R
\name{gargle-options}
\alias{gargle-options}
\alias{gargle_quiet}
\title{Options consulted by gargle}
\usage{
gargle_quiet()
}
\description{
Options consulted by gargle
}
\section{\code{gargle_quiet}}{

\code{gargle_quiet()} returns the option named "gargle_quiet", which defaults to
\code{TRUE}. Set this option to \code{FALSE} to see more info about gargle's
activities, which can be helpful for troubleshooting.
}

