% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gargle-class.R
\name{gargle2.0_token}
\alias{gargle2.0_token}
\title{Generate a gargle token}
\usage{
gargle2.0_token(email = getOption("gargle.oauth_email"),
  app = gargle_app(), package = "gargle", scope = NULL,
  user_params = NULL, type = NULL,
  use_oob = getOption("gargle.oob_default"), credentials = NULL,
  cache = if (is.null(credentials)) getOption("gargle.oauth_cache") else
  FALSE, ...)
}
\arguments{
\item{email}{Optional. Allows user to target a specific Google identity. If
specified, this is used for token lookup, i.e. to determine if a suitable
token is already available in the cache. If no such token is found, \code{email}
is used to pre-select the targetted Google identity in the OAuth chooser.
Note, however, that the email associated with a token when it's cached is
always determined from the token itself, never from this argument. Use \code{NA}
or \code{FALSE} to match nothing and force the OAuth dance in the browser. Use
\code{TRUE} to allow email auto-discovery, if a suitable token is found in the
cache. Define the option \code{gargle.oauth_email} to set a personal default.}

\item{app}{An OAuth consumer application, created by \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{package}{Name of the package requesting a token. Used in messages.}

\item{scope}{A character vector of scopes to request. The \code{"email"} scope is
always added if not already present. This is needed to retrieve the email
address associated with the token. This is considered a low value scope and
does not appear on the consent screen.}

\item{user_params}{Named list holding endpoint specific parameters to pass to
the server when posting the request for obtaining or refreshing the
access token.}

\item{type}{content type used to override incorrect server response}

\item{use_oob}{If \code{FALSE}, use a local webserver for the OAuth dance.
Otherwise, provide a URL to the user and prompt for a validation code.
Defaults to the option \code{gargle.oob_default} or \code{TRUE} if httpuv is not
installed.}

\item{credentials}{Advanced use only: allows you to completely customise
token generation.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache using the
default user-level cache file, \code{~/.R/gargle/gargle-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics. A
string means use the specified path as the cache file.}

\item{...}{Absorbs arguments intended for use by non-OAuth2 credential
functions. Not used.}
}
\value{
An object of class \link{Gargle2.0}, either new or loaded from the cache.
}
\description{
Constructor function for objects of class \link{Gargle2.0}.
}
