% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-api-key.R
\name{gargle_api_key}
\alias{gargle_api_key}
\title{API key for demonstration purposes}
\description{
Some APIs accept unauthorized requests for public resources, in
which case the request must be sent with an API key in lieu of a token. This
function provides an API key for limited use in prototyping and for testing
and documentation of gargle itself. This key may be deleted or rotated at any
time. This function is not exported and will only work inside other gargle
functions, such as \code{\link[=request_build]{request_build()}}.

To get your own API key to without these restrictions, setup a new Google
Cloud Platform project in \href{https://console.developers.google.com}{Google Developers Console}, enable the APIs of interest,
and follow the instructions in \href{https://support.google.com/googleapi/answer/6158862}{Setting up API keys}.
}
\examples{
\dontrun{
# use an API key with the Places API (explicitly enabled for this key)
# gets restaurants close to a location in Vancouver, BC
req <- request_build(
  method = "GET",
  path = "maps/api/place/nearbysearch/json",
  params = list(
    location = "49.268682,-123.167117",
    radius = 100,
    type = "restaurant"
  ),
  key = gargle:::gargle_api_key(),
  base_url = "https://maps.googleapis.com"
)
resp <- request_make(req)
out <- response_process(resp)
vapply(out$results, function(x) x$name, character(1))
}
}
\keyword{internal}
