% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-oauth-app.R
\name{gargle_app}
\alias{gargle_app}
\title{OAuth app for demonstration purposes}
\usage{
gargle_app()
}
\value{
An OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}},
invisibly.
}
\description{
Invisibly returns an OAuth app that can be used to test drive
gargle before obtaining your own client ID and secret. This OAuth app may
be deleted or rotated at any time. There are no guarantees about which APIs
are enabled. DO NOT USE THIS IN A PACKAGE or for anything other than
interactive, small-scale experimentation.

To get your own OAuth app, without these limitations, set up a new Google
\href{https://support.google.com/cloud/answer/6158853}{Cloud Platform Project}
in \href{https://console.developers.google.com}{Google Developers Console} and
complete the
\href{https://developers.google.com/identity/protocols/OAuth2InstalledApp#prerequisites}{Prerequisites}
for OAuth 2.0 for Mobile & Desktop Apps. Use \code{\link[httr:oauth_app]{httr::oauth_app()}} to create
an object of type \code{oauth_app}, suitable for use with the gargle and httr
packages.
}
\examples{
\dontrun{
gargle_app()
}
}
