% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{gargle_env}
\alias{gargle_env}
\title{Environment used for gargle global state.}
\format{An environment.}
\usage{
gargle_env
}
\description{
This environment contains:
\itemize{
\item \code{$cred_funs}: an ordered list of credential methods to use when trying
to fetch credentials.
\item \code{$auth}: a list that records current auth state:
\itemize{
\item \code{$active} Logical, indicates whether a token or \code{NULL} should be
delivered upon request via \code{token_deliver()}.
\item \code{$token} The most recently fetched token. If none has been fetched, value
is \code{NULL}.
\item \code{$method} The name of the credential function that produced the token. If
none has been used, value if \code{NA_character_}.
}
}
}
\keyword{internal}
