% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{gargle_map_cli}
\alias{gargle_map_cli}
\title{Map a cli-styled template over an object}
\usage{
gargle_map_cli(x, ...)
}
\description{
For internal use in gargle, googledrive, and googlesheets4 (for now).
}
\keyword{internal}
