% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-cache.R
\name{gargle_oauth_sitrep}
\alias{gargle_oauth_sitrep}
\title{OAuth token situation report}
\usage{
gargle_oauth_sitrep(cache = NULL)
}
\arguments{
\item{cache}{Specifies the OAuth token cache. Defaults to the option named
\code{"gargle_oauth_cache"}, retrieved via \code{\link[=gargle_oauth_cache]{gargle_oauth_cache()}}.}
}
\value{
A data frame with one row per cached token, invisibly. Note this data
frame contains more columns than it seems, e.g. the \code{filepath} column
isn't printed by default.
}
\description{
Get a human-oriented overview of the existing gargle OAuth tokens:
\itemize{
\item Filepath of the current cache
\item Number of tokens found there
\item Compact summary of the associated
\itemize{
\item Email = Google identity
\item OAuth client (actually, just its nickname)
\item Scopes
\item Hash (actually, just the first 7 characters)
}
}

This is the most direct way to find out where gargle's OAuth cache is (see
\code{\link[=gargle_oauth_cache]{gargle_oauth_cache()}} for more about the cache location). The \code{filepath}
column of the returned data frame makes it easy to delete specific tokens,
e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{dat <- gargle_oauth_sitrep()
# delete all cached tokens for one Google identity
unlink(dat$filepath[dat$email == "jane@example.com"])
}\if{html}{\out{</div>}}
}
\examples{
gargle_oauth_sitrep()
}
