% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_gce.R
\name{gce_instance_service_accounts}
\alias{gce_instance_service_accounts}
\title{List all service accounts available on this GCE instance}
\usage{
gce_instance_service_accounts()
}
\value{
A data frame, where each row is a service account. Due to aliasing,
there is no guarantee that each row represents a distinct service account.
}
\description{
List all service accounts available on this GCE instance
}
\examples{
\dontshow{if (gargle:::is_gce()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
credentials_gce()
\dontshow{\}) # examplesIf}
}
\seealso{
The return value is built from a recursive query of the so-called
"directory" of the instance's service accounts as documented in
\url{https://cloud.google.com/compute/docs/metadata/default-metadata-values#vm_instance_metadata}.
}
