% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application-default-credentials.R
\name{get_application_default_credentials}
\alias{get_application_default_credentials}
\title{Fetch the ADC.}
\usage{
get_application_default_credentials(scopes, ...)
}
\description{
Fetch the ADC.
}
