% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce-token.R
\name{get_gce_credentials}
\alias{get_gce_credentials}
\title{Create a token for use on Google Compute Engine for the given scopes, if possible.}
\usage{
get_gce_credentials(scopes, service_account = "default", ...)
}
\arguments{
\item{scopes}{List of scopes required for the returned token.}

\item{service_account}{Name of the GCE service account to use (defaults to \code{default})}

\item{...}{Additional arguments passed to all credentials functions.}
}
\value{
A \code{\link[=GceToken]{GceToken()}} or \code{NULL}.
}
\description{
Create a token for use on Google Compute Engine for the given scopes, if possible.
}
