% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service-account.R
\name{get_service_account_credentials}
\alias{get_service_account_credentials}
\title{Create a token for a Google service account.}
\usage{
get_service_account_credentials(scopes, path = "", ...)
}
\arguments{
\item{scopes}{List of scopes required for the returned token.}

\item{path}{Path to the downloaded JSON file}
}
\value{
A \code{httr::TokenServiceAccount} or \code{NULL}.
}
\description{
Create a token for a Google service account.
}
