% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-email.R
\name{get_userinfo}
\alias{get_userinfo}
\alias{get_email}
\title{Get info on current user}
\usage{
get_userinfo(token)

get_email(token)
}
\arguments{
\item{token}{A gargle token}
}
\value{
User info or user's email
}
\description{
The userinfo endpoint can potentially return the user's entire profile, if
the token has the necessary scope. But, by default, gargle-mediated tokens
only include the \code{userinfo.email} scope, so don't expect to get much beyond
the email, unless you've explicitly scoped for more. \code{get_userinfo()} can
exploit built-in token refresh, although at the time of writing, the
motivating use case is to get and store the email for a freshly-obtained
token.
}
\examples{
\dontrun{
# with service account token
t <- token_fetch(
  scopes = "https://www.googleapis.com/auth/drive",
  path   = "path/to/service/account/token/blah-blah-blah.json"
)
# or with an OAuth token
t <- token_fetch(
  scopes = "https://www.googleapis.com/auth/drive",
  email  = "janedoe@example.com"
)
get_userinfo(t)
get_email(t)
get_tokeninfo(t)
}
}
\keyword{internal}
