% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle_api_key.R, R/gargle_oauth_client.R
\name{internal-assets}
\alias{internal-assets}
\alias{tidyverse_api_key}
\alias{tidyverse_client}
\alias{tidyverse_app}
\title{Assets for internal use}
\usage{
tidyverse_api_key()

tidyverse_client(type = NULL)

tidyverse_app()
}
\description{
Assets for use inside specific packages maintained by the tidyverse team.
}
\keyword{internal}
