% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{gargle_last_response}
\alias{gargle_last_response}
\alias{gargle_last_content}
\title{Access the last response}
\usage{
gargle_last_response()

gargle_last_content()
}
\value{
\itemize{
\item \code{gargle_last_response()} returns the most recent \code{\link[httr:response]{httr::response()}}
object.
\item \code{gargle_last_content()} returns the parsed JSON content from the most
recent response or an empty list if unavailable.
}
}
\description{
These functions give access to the most recent response processed by
\code{\link[=response_process]{response_process()}} (if the argument \code{remember = TRUE}, which is the
default). They can be useful for \emph{post mortem} analysis of puzzling or
failed API interactions.
}
\keyword{internal}
