% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_gce.R
\name{list_service_accounts}
\alias{list_service_accounts}
\title{List all service accounts available on this GCE instance.}
\usage{
list_service_accounts()
}
\value{
A list of service account names.
}
\description{
List all service accounts available on this GCE instance.
}
