% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-factories.R
\name{make_auth}
\alias{make_auth}
\title{Authorization function factory}
\usage{
make_auth(pkg_name, scopes,
  error_message = "Can't get Google credentials.\\nAre you running this package in a non-interactive session? Consider calling this authorization function directly with all necessary specifics.\\n")
}
\arguments{
\item{pkg_name}{The name of your package.}

\item{scopes}{A character vector of the Google api scopes needed by your
package.}

\item{error_message}{The error message to show users when authorization
fails.}
}
\value{
An authorization function.
}
\description{
Create a function in your package for Google authorization.
}
\examples{
drive_auth <- make_auth(
  scopes = "https://www.googleapis.com/auth/drive",
  error_message <- "Can't get Google credentials.\\\\nAre you running googledrive in a non-interactive session? Consider:\\\\n  * `drive_deauth()` to prevent the attempt to get credentials.\\\\n  * Call `drive_auth()` directly with all necessary specifics.\\\\n"
)
}
