% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-factories.R
\name{make_deauth}
\alias{make_deauth}
\title{Deauthorization function factory}
\usage{
make_deauth()
}
\value{
A deauthorization function.
}
\description{
Create a function in your package for Google deauthorization.
}
\examples{
drive_deauth <- make_deauth()
}
