% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_github_actions.R
\name{oauth_gha_token}
\alias{oauth_gha_token}
\title{Generate OAuth token for an external account on Github Actions}
\usage{
oauth_gha_token(
  project_id,
  workload_identity_provider,
  service_account,
  lifetime,
  scopes = "https://www.googleapis.com/auth/drive.file",
  id_token_url = Sys.getenv("ACTIONS_ID_TOKEN_REQUEST_URL"),
  id_token_request_token = Sys.getenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN")
)
}
\arguments{
\item{project_id}{The google cloud project id}

\item{workload_identity_provider}{The workload identity provider}

\item{service_account}{The service account email address}

\item{lifetime}{Lifespan of token in seconds as a string \code{"300s"}}

\item{scopes}{Requested scopes for the access token}

\item{universe}{Set the domain for the endpoints}
}
\description{
Generate OAuth token for an external account on Github Actions
}
\keyword{internal}
