% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{require_oauth}
\alias{require_oauth}
\title{Require OAuth login for Shiny app}
\usage{
require_oauth(
  app,
  oauth_app,
  scopes,
  welcome_ui,
  cookie_opts = cookie_options(http_only = TRUE)
)
}
\arguments{
\item{app}{The return value from \code{\link[shiny:shinyApp]{shiny::shinyApp()}}. For readability,
consider using a pipe operator, i.e. \code{shinyApp() \%>\% require_oauth(...)}.}

\item{oauth_app}{An \code{\link[httr:oauth_app]{httr::oauth_app()}} object that provides the OAuth client
ID and secret. See the \href{https://gargle.r-lib.org/articles/get-api-credentials.html}{How to get your own API credentials}
vignette and \code{\link[=oauth_app_from_json]{oauth_app_from_json()}}.}

\item{scopes}{A character vector of scopes to request. Pick from those listed
at \url{https://developers.google.com/identity/protocols/googlescopes}.

For certain token flows, the
\code{"https://www.googleapis.com/auth/userinfo.email"} scope is unconditionally
included. This grants permission to retrieve the email address associated
with a token; gargle uses this to index cached OAuth tokens. This grants no
permission to view or send email. It is considered a low value scope and
does not appear on the consent screen.}
}
\description{
Use this function to enforce Google Auth login for all visitors to a Shiny
app. Once logged in, a \link[=Gargle-class]{token} will be stored on the Shiny
session object and automatically used for any Google API operations that go
through gargle.
}
