% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{set_credential_functions}
\alias{set_credential_functions}
\title{Set the list of all credential functions.}
\usage{
set_credential_functions(ls)
}
\arguments{
\item{ls}{A list of credential functions.}
}
\description{
Set the list of all credential functions.
}
\seealso{
Other registration: \code{\link{add_credential_function}},
  \code{\link{all_credential_functions}}
}
