% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{set_default_credential_functions}
\alias{set_default_credential_functions}
\title{Set the default credential functions.}
\usage{
set_default_credential_functions()
}
\description{
Set the default credential functions.
}
