% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-api-key.R
\name{tidyverse_api_key}
\alias{tidyverse_api_key}
\title{API key for tidyverse packages}
\usage{
tidyverse_api_key()
}
\value{
A Google API key
}
\description{
Returns an API key for use in tidyverse packages, e.g., googledrive,
googlesheets, bigrquery. Please don't use this API key in non-tidyverse
projects. For a default API key to use while getting to know gargle and
instructions on how to get your own key, see \code{\link[=gargle_api_key]{gargle_api_key()}}.
}
\examples{
tidyverse_api_key()
}
\keyword{internal}
