% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-oauth-app.R
\name{tidyverse_app}
\alias{tidyverse_app}
\title{Oauth app for tidyverse packages}
\usage{
tidyverse_app()
}
\value{
An OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
Returns an OAuth app for use in tidyverse packages, e.g., googledrive,
googlesheets, bigrquery. Developers of other packages should create their own
\href{https://support.google.com/cloud/answer/6158853}{Cloud Platform Project} in
the \href{https://console.developers.google.com}{Google Developers Console} and
obtain their own client ID and secret. These can then be used with
\code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\examples{
tidyverse_app()
}
