% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-fetch.R
\name{token_deliver}
\alias{token_deliver}
\title{Hand over a token}
\usage{
token_deliver(scopes = "https://www.googleapis.com/auth/drive", ...)
}
\arguments{
\item{scopes}{scopes!}

\item{...}{dots!}
}
\value{
things!
}
\description{
Coughs up a token. From gargle's cache, if possible. Fetches new one if
necessary. Returns \code{NULL} if auth is inactive.
}
\examples{
\dontrun{
token_deliver()
}
}
