% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-fetch.R
\name{token_ingest}
\alias{token_ingest}
\title{Ingest a token}
\usage{
token_ingest(token)
}
\arguments{
\item{token}{Path to an \code{.rds} file containing a token or an actual token.}
}
\value{
logical
}
\description{
Ingest a token
}
