% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{add_credential_function}
\alias{add_credential_function}
\title{Add a new credential fetching function.}
\usage{
add_credential_function(...)
}
\arguments{
\item{...}{One or more functions with the right signature. See
\code{\link[=is_credential_function]{is_credential_function()}}.}
}
\description{
Note that this implicitly adds \code{f} to the \emph{front} of the list.
}
\examples{
creds_one <- function(scopes, ...) {}
add_credential_function(creds_one)
add_credential_function(one = creds_one)
add_credential_function(one = creds_one, two = creds_one)
add_credential_function(one = creds_one, creds_one)
}
\seealso{
Other registration: \code{\link{all_credential_functions}},
  \code{\link{clear_credential_functions}},
  \code{\link{set_credential_functions}}
}
