% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-state.R
\name{auth_activate}
\alias{auth_activate}
\alias{auth_deactivate}
\alias{auth_state}
\title{Activate / de-activate auth}
\usage{
auth_activate(verbose = TRUE)

auth_deactivate(clear = FALSE, verbose = TRUE)

auth_state(verbose = TRUE)
}
\arguments{
\item{verbose}{verbosity!}

\item{clear}{Logical, whether to clear the cached token}
}
\value{
the auth state, invisibly
}
\description{
Activate / de-activate auth
}
