% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthState-class.R
\docType{class}
\name{AuthState-class}
\alias{AuthState-class}
\alias{AuthState}
\title{Authorization state}
\format{An R6 class object.}
\usage{
AuthState
}
\arguments{
\item{app}{An OAuth consumer application, created by \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{api_key}{API key. Necessary in order to make unauthorized "token-free"
requests for public resources. Can be set to \code{NULL} if all requests will be
authorized, i.e. they will include a token.}

\item{auth_active}{Logical. \code{TRUE} means requests should include a token (and
not an API key). \code{FALSE} means requests should include an API key (and not
a token).}

\item{cred}{A configured token.}
}
\description{
An \code{AuthState} object manages an authorization state, presumably on behalf of
a client package that makes requests to a Google API. This state is
incorporated into the package's requests for tokens and controls the
inclusion of tokens in requests to the target API:
\itemize{
\item \code{app} and \code{api_key} identify the package to Google APIs as an
"application".
\item \code{auth_active} reflects whether requests are authorized by an
authenticated user or are unauthorized requests for public resources.
\item \code{cred} is a configured token, ready to send in requests. If
\code{auth_active = FALSE}, this should be \code{NULL}.
}
}
\keyword{internal}
