% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-function-registry.R
\name{cred_funs}
\alias{cred_funs}
\alias{cred_funs_list}
\alias{cred_funs_add}
\alias{cred_funs_set}
\alias{cred_funs_clear}
\alias{cred_funs_set_default}
\title{Credential function registry}
\usage{
cred_funs_list()

cred_funs_add(...)

cred_funs_set(ls)

cred_funs_clear()

cred_funs_set_default()
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more credential
functions, in \code{name = value} form. Each credential function is subject to a
superficial check that it at least "smells like" a credential function: its
first argument must be named \code{scopes}, and its signature must include
\code{...}. To remove a credential function, you can use a specification like
\code{name = NULL}.}

\item{ls}{A named list of credential functions.}
}
\value{
A list of credential functions or \code{NULL}.
}
\description{
Functions to query or manipulate the registry of credential functions
consulted by \code{\link[=token_fetch]{token_fetch()}}.
}
\section{Functions}{
\itemize{
\item \code{cred_funs_list()}: Get the list of registered credential functions.

\item \code{cred_funs_add()}: Register one or more new credential fetching functions.
Function(s) are added to the \emph{front} of the list. So:
\itemize{
\item "First registered, last tried."
\item "Last registered, first tried."
}

\item \code{cred_funs_set()}: Register a list of credential fetching functions.

\item \code{cred_funs_clear()}: Clear the credential function registry.

\item \code{cred_funs_set_default()}: Reset the registry to the gargle default.

}}
\examples{
names(cred_funs_list())

creds_one <- function(scopes, ...) {}

cred_funs_add(one = creds_one)
cred_funs_add(two = creds_one, three = creds_one)
names(cred_funs_list())

cred_funs_add(two = NULL)
names(cred_funs_list())

# undo all of the above and return to default
cred_funs_set_default()
}
\seealso{
\code{\link[=token_fetch]{token_fetch()}}, which is where the registry is actually used.
}
