% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cred_funs_clear}
\alias{cred_funs_clear}
\title{Clear the list of credential functions.}
\usage{
cred_funs_clear()
}
\description{
Clear the list of credential functions.
}
\seealso{
Other registration: \code{\link{cred_funs_add}},
  \code{\link{cred_funs_list}}, \code{\link{cred_funs_set}}
}
\concept{registration}
