% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{cred_funs_set}
\alias{cred_funs_set}
\title{Set the list of all credential functions.}
\usage{
cred_funs_set(ls)
}
\arguments{
\item{ls}{A list of credential functions.}
}
\description{
Set the list of all credential functions.
}
\seealso{
Other registration: \code{\link{cred_funs_add}},
  \code{\link{cred_funs_clear}},
  \code{\link{cred_funs_list}}
}
