% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application-default-credentials.R
\name{credentials_app_default}
\alias{credentials_app_default}
\title{Fetch the Application Default Credentials.}
\usage{
credentials_app_default(scopes, ...)
}
\arguments{
\item{scopes}{List of scopes required for the returned token.}

\item{...}{Additional arguments (ignored)}
}
\description{
Fetch the Application Default Credentials.
}
