% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_github_actions.R
\name{credentials_github_actions}
\alias{credentials_github_actions}
\title{Get a token using Github Actions}
\usage{
credentials_github_actions(
  project_id,
  workload_identity_provider,
  service_account,
  lifetime = "300s",
  scopes = "https://www.googleapis.com/auth/drive.file",
  ...
)
}
\arguments{
\item{project_id}{The google cloud project id}

\item{workload_identity_provider}{The workload identity provider}

\item{service_account}{The service account email address}

\item{lifetime}{Lifespan of token in seconds as a string \code{"300s"}}

\item{scopes}{Requested scopes for the access token}

\item{...}{Additional arguments passed to all credential functions.}
}
\value{
A \code{\link[=WifToken]{WifToken()}} or \code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
credentials_github_actions(
  project_id = "project-id-12345",
  workload_identity_provider = "projects/123456789/locations/global/workloadIdentityPools/my-pool/providers/my-provider"
  service_account = "my-service-account@my-project.iam.gserviceaccount.com",
  scopes = "https://www.googleapis.com/auth/drive.file"
)
}
}
\seealso{
There is some setup required in GCP to enable this auth flow.
This function reimplements the \code{google-github-actions/auth}. The
documentation for that workflow provides instructions on the setup steps.
\itemize{
\item \url{https://github.com/google-github-actions/auth?tab=readme-ov-file#indirect-wif}
}

Other credential functions: 
\code{\link{credentials_app_default}()},
\code{\link{credentials_byo_oauth2}()},
\code{\link{credentials_external_account}()},
\code{\link{credentials_gce}()},
\code{\link{credentials_service_account}()},
\code{\link{credentials_user_oauth2}()},
\code{\link{token_fetch}()}
}
\concept{credential functions}
