% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-legged-oauth.R
\name{credentials_user_oauth2}
\alias{credentials_user_oauth2}
\title{Create a token for a user via the browser flow.}
\usage{
credentials_user_oauth2(scopes, oauth_app = NULL, ...)
}
\arguments{
\item{scopes}{List of scopes required for the returned token.}

\item{oauth_app}{\code{httr::oauth_app} to use for this token fetch. (Optional.)}

\item{...}{Additional arguments (ignored)}
}
\description{
Create a token for a user via the browser flow.
}
