% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-legged-oauth.R
\name{credentials_user_oauth2}
\alias{credentials_user_oauth2}
\title{Create a token for a user via the browser flow.}
\usage{
credentials_user_oauth2(scopes, app, cache = getOption("httr_oauth_cache"),
  ...)
}
\arguments{
\item{scopes}{A character vector of scopes to request.}

\item{app}{An OAuth consumer application, created by \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache using the
default cache file \code{.httr-oauth}. \code{FALSE} means don't cache, and \code{NA} means
to guess using some sensible heuristics. A string means use the specified
path as the cache file.}

\item{...}{Additional arguments (ignored)}
}
\description{
Create a token for a user via the browser flow.
}
