% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gargle-class.R
\docType{class}
\name{Gargle-class}
\alias{Gargle-class}
\alias{Gargle2.0}
\title{OAuth2 token objects specific to Google APIs}
\format{An R6 class object.}
\usage{
Gargle2.0
}
\description{
This is based on the \link[httr:Token-class]{Token2.0} class provided in httr.
These objects should be created through the constructor function
\code{\link[=gargle2.0_token]{gargle2.0_token()}}. In the base Token2.0 class, tokens are cached based on
endpoint, app, and scopes. For the \code{Gargle2.0} subclass, the identifier or
key is expanded to include the email address associated with the token. This
makes it easier to work with Google APIs with multiple identities. The
default location for the token cache is also different: it's now
\code{"~/.R/gargle/gargle-oauth"} instead of \code{./.httr-oauth}.
}
\keyword{internal}
