% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-package.R
\docType{package}
\name{gargle-package}
\alias{gargle}
\alias{gargle-package}
\title{gargle: Utilities for Working with Google APIs}
\description{
Provides utilities for working with Google APIs \url{https://developers.google.com/apis-explorer}. This includes functions and classes for handling common credential types and for preparing, executing, and processing HTTP requests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gargle.r-lib.org}
  \item \url{https://github.com/r-lib/gargle}
  \item Report bugs at \url{https://github.com/r-lib/gargle/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Craig Citro \email{craigcitro@google.com}
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item Google Inc [copyright holder]
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
