% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-package.R
\docType{package}
\name{gargle-package}
\alias{gargle}
\alias{gargle-package}
\title{gargle: Easier Access to Google Services from R}
\description{
Utilities to take some of the pain out of working
    with Google services and APIs. This includes functions and classes for
    handling common credential types and for streamlining the preparation
    of HTTP requests, based on the API specs provided in its Discovery
    Document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/gargle}
  \item Report bugs at \url{https://github.com/r-lib/gargle/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com} (0000-0002-6983-2759)

Authors:
\itemize{
  \item Craig Citro \email{craigcitro@google.com}
}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} (0000-0003-4757-117X) [contributor]
  \item Google Inc [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
