% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-package.R
\docType{package}
\name{gargle-package}
\alias{gargle}
\alias{gargle-package}
\title{gargle: Utilities for Working with Google APIs from R}
\description{
Utilities to take some of the pain out of working
    with Google APIs.  This includes functions and classes for handling
    common credential types and for preparing, executing, and processing
    HTTP requests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gargle.r-lib.org}
  \item \url{https://github.com/r-lib/gargle}
  \item Report bugs at \url{https://github.com/r-lib/gargle/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com} (0000-0002-6983-2759)

Authors:
\itemize{
  \item Craig Citro \email{craigcitro@google.com}
  \item Hadley Wickham \email{hadley@rstudio.com} (0000-0003-4757-117X)
}

Other contributors:
\itemize{
  \item Google Inc [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
