% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gargle-class.R
\name{gargle2.0_token}
\alias{gargle2.0_token}
\title{Generate a Gargle token}
\usage{
gargle2.0_token(email = NULL, app = gargle_app(), scope = NULL,
  user_params = NULL, type = NULL,
  use_oob = getOption("httr_oob_default"), credentials = NULL,
  cache = getOption("httr_oauth_cache"))
}
\arguments{
\item{email}{Email address}

\item{app}{An OAuth consumer application, created by
\code{\link{oauth_app}}}

\item{scope}{a character vector of scopes to request.}

\item{user_params}{Named list holding endpoint specific parameters to pass to
the server when posting the request for obtaining or refreshing the
access token.}

\item{type}{content type used to override incorrect server response}

\item{use_oob}{if FALSE, use a local webserver for the OAuth dance.
Otherwise, provide a URL to the user and prompt for a validation
code. Defaults to the of the \code{"httr_oob_default"} default,
or \code{TRUE} if \code{httpuv} is not installed.}

\item{credentials}{Advanced use only: allows you to completely customise
token generation.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string mean use the specified path as the cache file.}
}
\value{
A Gargle token
}
\description{
Generate a Gargle token
}
