% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gargle-class.R
\name{gargle2.0_token}
\alias{gargle2.0_token}
\title{Generate a Gargle token}
\usage{
gargle2.0_token(email = NULL, app = gargle_app(), scope = NULL,
  user_params = NULL, type = NULL,
  use_oob = getOption("gargle.oob_default"), credentials = NULL,
  cache = if (is.null(credentials)) getOption("gargle.oob_default") else
  FALSE, ...)
}
\arguments{
\item{email}{Optional. Allows user to target a specific Google identity. If
specified, this is used only for token lookup, i.e. to determine if a
suitable token is already available in the cache. The email associated with
a token when it's cached is determined from the token itself, not from this
argument.}

\item{app}{An OAuth consumer application, created by
\code{\link{oauth_app}}}

\item{scope}{a character vector of scopes to request.}

\item{user_params}{Named list holding endpoint specific parameters to pass to
the server when posting the request for obtaining or refreshing the
access token.}

\item{type}{content type used to override incorrect server response}

\item{use_oob}{if FALSE, use a local webserver for the OAuth dance.
Otherwise, provide a URL to the user and prompt for a validation
code. Defaults to the of the \code{"httr_oob_default"} default,
or \code{TRUE} if \code{httpuv} is not installed.}

\item{credentials}{Advanced use only: allows you to completely customise
token generation.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string mean use the specified path as the cache file.}

\item{...}{Absorbs arguments intended for use by non-OAuth2 credential
functions. Not used.}
}
\value{
An object of class \link{Gargle2.0}, either new or loaded from the
cache.
}
\description{
Constructor function for objects of class \link{Gargle2.0}. The \code{"email"} scope
is always added if not already present. This is needed to retrieve the email
address associated with the token. This is considered a "low value" scope and
does not appear on the consent screen.
}
