% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-api-key.R
\name{gargle_api_key}
\alias{gargle_api_key}
\title{API key for demonstration purposes}
\usage{
gargle_api_key()
}
\description{
Some APIs accept requests for public resources, in which case
the request must be sent with an API key in lieu of a token. This function
provides an API key for limited use in prototyping and for testing and
documentation of gargle itself. This key may be deleted or rotated at any
time. There are no guarantees about which APIs are enabled. DO NOT USE THIS
IN A PACKAGE or for anything other than interactive, small-scale
experimentation.

You can get your own API key, without these limitations. See the \href{https://gargle.r-lib.org/articles/get-api-credentials.html}{How to get your own API credentials}
vignette for more details.
}
\examples{
gargle_api_key()
}
\keyword{internal}
