% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-api-key.R
\name{gargle_api_key}
\alias{gargle_api_key}
\title{API key for demonstration purposes}
\usage{
gargle_api_key()
}
\description{
Some APIs accept requests for public resources, in which case
the request must be sent with an API key in lieu of a token. This function
provides an API key for limited use in prototyping and for testing and
documentation of gargle itself. This key may be deleted or rotated at any
time. There are no guarantees about which APIs are enabled. DO NOT USE THIS
IN A PACKAGE or for anything other than interactive, small-scale
experimentation.

To get your own API key, without these limitations, set up a new Google
Cloud Platform project in \href{https://console.developers.google.com}{Google Developers Console}, enable the APIs of
interest, and follow the instructions in \href{https://support.google.com/googleapi/answer/6158862}{Setting up API keys}.
}
\examples{
\dontrun{
# use an API key for which the Places API is enabled!
API_KEY <- "1234567890"

# get restaurants close to a location in Vancouver, BC
req <- request_build(
  method = "GET",
  path = "maps/api/place/nearbysearch/json",
  params = list(
    location = "49.268682,-123.167117",
    radius = 100,
    type = "restaurant"
  ),
  key = API_KEY,
  base_url = "https://maps.googleapis.com"
)
resp <- request_make(req)
out <- response_process(resp)
vapply(out$results, function(x) x$name, character(1))
}
}
\keyword{internal}
