% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-oauth-app.R
\name{gargle_app}
\alias{gargle_app}
\title{OAuth app for demonstration purposes}
\usage{
gargle_app()
}
\value{
An OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}},
invisibly.
}
\description{
Invisibly returns an OAuth app that can be used to test drive
gargle before obtaining your own client ID and secret. This OAuth app may
be deleted or rotated at any time. There are no guarantees about which APIs
are enabled. DO NOT USE THIS IN A PACKAGE or for anything other than
interactive, small-scale experimentation.

You can get your own OAuth app (client ID and secret), without these
limitations. See the \href{https://gargle.r-lib.org/articles/get-api-credentials.html}{How to get your own API credentials}
vignette for more details.
}
\examples{
\dontrun{
gargle_app()
}
}
\keyword{internal}
