% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle-oauth-app.R
\name{gargle_app}
\alias{gargle_app}
\title{OAuth app for demonstration purposes}
\usage{
gargle_app()
}
\value{
An OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
Returns an OAuth app that we default to in \code{\link[=credentials_user_oauth2]{credentials_user_oauth2()}}. This
makes it easy to test drive gargle before obtaining your own client ID and
secret. Once your use of gargle is more than casual, you should create a
similar function, but populated with your own credentials, like so:
\itemize{
\item Create a \href{https://support.google.com/cloud/answer/6158853}{Cloud Platform Project}
in the \href{https://console.developers.google.com}{Google Developers Console}.
\item Create a function, e.g., \code{my_app()}, that calls \code{\link[httr:oauth_app]{httr::oauth_app()}}
with a meaningful nickname and your app's client ID and secret.
}
}
\examples{
\dontrun{
gargle_app()
}
}
