% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{gargle_env}
\alias{gargle_env}
\title{Environment used for gargle global state.}
\format{An environment.}
\usage{
gargle_env
}
\description{
This environment contains:
\itemize{
\item \code{$cred_funs} is the ordered list of credential methods to use when trying
to fetch credentials.
}
}
\keyword{internal}
