% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-cache.R
\name{gargle_oauth_sitrep}
\alias{gargle_oauth_sitrep}
\title{OAuth token situation report}
\usage{
gargle_oauth_sitrep(cache = NULL)
}
\arguments{
\item{cache}{Specifies the OAuth token cache. Defaults to the option named
\code{"gargle_oauth_cache"}, retrieved via \code{\link[=gargle_oauth_cache]{gargle_oauth_cache()}}.}
}
\value{
A data frame with one row per cached token, invisibly. Note this data
frame may contain more columns than it seems, e.g. the \code{filepath} column
isn't printed by default.
}
\description{
Get a human-oriented overview of the existing gargle OAuth tokens:
\itemize{
\item Filepath of the current cache
\item Number of tokens found there
\item Compact summary of the associated
\itemize{
\item Email = Google identity
\item OAuth app (actually, just its nickname)
\item Scopes
\item Hash (actually, just the first 7 characters)
Mostly useful for the development of gargle and client packages.
}
}
}
\examples{
gargle_oauth_sitrep()
}
