% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-cache.R
\name{gargle_oauth_sitrep}
\alias{gargle_oauth_sitrep}
\title{OAuth token situation report}
\usage{
gargle_oauth_sitrep(cache = NULL)
}
\arguments{
\item{cache}{A logical value or a string. \code{TRUE} means to cache using the
default user-level cache file, \code{~/.R/gargle/gargle-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics. A
string means use the specified path as the cache file.}
}
\value{
A data frame with one row per cached token, invisibly.
}
\description{
Get a human-oriented overview of the existing gargle OAuth tokens:
\itemize{
\item Filepath of the current cache
\item Number of tokens found there
\item Compact summary of the associated
\itemize{
\item Email = Google identity
\item OAuth app (actually, just its nickname)
\item Scopes
\item Hash (actually, just the first 7 characters)
Mostly useful for the development of gargle and client packages.
}
}
}
\examples{
gargle_oauth_sitrep()
}
