% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_gce.R
\name{GceToken}
\alias{GceToken}
\title{Token for use on Google Compute Engine instances}
\description{
This class uses the metadata service available on GCE VMs to fetch access
tokens. Not intended for direct use. See \code{\link[=credentials_gce]{credentials_gce()}} instead.
}
\section{Super classes}{
\code{\link[httr:Token]{httr::Token}} -> \code{\link[httr:Token2.0]{httr::Token2.0}} -> \code{GceToken}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{GceToken$print()}}
\item \href{#method-init_credentials}{\code{GceToken$init_credentials()}}
\item \href{#method-cache}{\code{GceToken$cache()}}
\item \href{#method-load_from_cache}{\code{GceToken$load_from_cache()}}
\item \href{#method-can_refresh}{\code{GceToken$can_refresh()}}
\item \href{#method-refresh}{\code{GceToken$refresh()}}
\item \href{#method-revoke}{\code{GceToken$revoke()}}
\item \href{#method-clone}{\code{GceToken$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="hash">}\href{../../httr/html/Token.html#method-hash}{\code{httr::Token$hash()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="initialize">}\href{../../httr/html/Token.html#method-initialize}{\code{httr::Token$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token2.0" data-id="sign">}\href{../../httr/html/Token2.0.html#method-sign}{\code{httr::Token2.0$sign()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token2.0" data-id="validate">}\href{../../httr/html/Token2.0.html#method-validate}{\code{httr::Token2.0$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_credentials"></a>}}
\subsection{Method \code{init_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$init_credentials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cache"></a>}}
\subsection{Method \code{cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$cache(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_from_cache"></a>}}
\subsection{Method \code{load_from_cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$load_from_cache(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-can_refresh"></a>}}
\subsection{Method \code{can_refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$can_refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\subsection{Method \code{refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-revoke"></a>}}
\subsection{Method \code{revoke()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$revoke()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GceToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
