% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_gce.R
\docType{data}
\name{GceToken}
\alias{GceToken}
\title{Token for use on Google Compute Engine instances}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GceToken
}
\description{
This class uses the metadata service available on GCE VMs to fetch access
tokens. Not intended for direct use. See \code{\link[=credentials_gce]{credentials_gce()}} instead.
}
\keyword{datasets}
