% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce-token.R
\docType{data}
\name{GceToken}
\alias{GceToken}
\title{GceToken is a token for use only on Google Compute Engine instances.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GceToken
}
\description{
This class uses the metadata service available on GCE VMs to fetch access tokens.
}
\keyword{datasets}
