% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application-default-credentials.R
\name{get_application_default_credentials}
\alias{get_application_default_credentials}
\title{Fetch the Application Default Credentials.}
\usage{
get_application_default_credentials(scopes, ...)
}
\arguments{
\item{scopes}{List of scopes required for the returned token.}
}
\description{
Fetch the Application Default Credentials.
}
