% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis-credentials.R
\name{get_travis_credentials}
\alias{get_travis_credentials}
\title{Create credentials from a service account token stored by travis.}
\usage{
get_travis_credentials(scopes, path = "", ...)
}
\description{
Create credentials from a service account token stored by travis.
}
