% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-legged-oauth.R
\name{get_user_oauth2_credentials}
\alias{get_user_oauth2_credentials}
\title{Create a token for a user via the browser flow.}
\usage{
get_user_oauth2_credentials(scopes, oauth_app = NULL, ...)
}
\description{
Create a token for a user via the browser flow.
}
