% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gargle_oauth_client.R
\name{oauth_app_from_json}
\alias{oauth_app_from_json}
\alias{gargle_app}
\title{Create an OAuth app from JSON}
\usage{
oauth_app_from_json(path, appname = NULL)

gargle_app()
}
\arguments{
\item{path}{JSON downloaded from \href{https://console.cloud.google.com}{Google Cloud Console}, containing a client id and
secret, in one of the forms supported for the \code{txt} argument of
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}

\item{appname}{name of the application.  This is not used for OAuth, but is
used to make it easier to identify different applications.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{oauth_app_from_json()} is being replaced with
\code{\link[=gargle_oauth_client_from_json]{gargle_oauth_client_from_json()}}, in light of the new
\code{gargle_oauth_client} class. Now \code{oauth_app_from_json()} potentially warns
about this deprecation and immediately passes its inputs through to
\code{\link[=gargle_oauth_client_from_json]{gargle_oauth_client_from_json()}}.

\code{gargle_app()} is being replaced with \code{\link[=gargle_client]{gargle_client()}}.
}
\keyword{internal}
