% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-app.R
\name{oauth_app_from_json}
\alias{oauth_app_from_json}
\title{Create an OAuth app from JSON}
\usage{
oauth_app_from_json(path, appname = NULL)
}
\arguments{
\item{path}{Path to the JSON file.}

\item{appname}{name of the application.  This is not used for OAuth, but
is used to make it easier to identify different applications and
provide a consistent way of storing secrets in environment variables.}
}
\description{
Essentially a wrapper around \code{\link[httr:oauth_app]{httr::oauth_app()}} that extracts client id (aka
key) and secret from JSON downloaded from \href{https://console.developers.google.com}{Google Developers Console}. If no \code{appname} is given,
the \code{"project_id"} from the JSON is used.
}
\examples{
\dontrun{
oauth_app(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}
}
