% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-oauth-app.R
\name{tidyverse_app}
\alias{tidyverse_app}
\title{OAuth app for tidyverse packages}
\usage{
tidyverse_app()
}
\value{
An OAuth consumer application, produced by \code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
Returns an OAuth app for use in tidyverse packages, e.g., googledrive,
googlesheets, bigrquery. Please don't use this app directly in non-tidyverse
projects.
}
\examples{
tidyverse_app()
}
\seealso{
For a default app to use while getting to know gargle and
instructions on how to make your own app, see \code{\link[=gargle_app]{gargle_app()}}.
}
\keyword{internal}
