% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_external_account.R
\name{WifToken}
\alias{WifToken}
\title{Token for use with workload identity federation}
\description{
Token for use with workload identity federation

Token for use with workload identity federation
}
\section{Super classes}{
\code{\link[httr:Token]{httr::Token}} -> \code{\link[httr:Token2.0]{httr::Token2.0}} -> \code{WifToken}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WifToken$new()}}
\item \href{#method-init_credentials}{\code{WifToken$init_credentials()}}
\item \href{#method-refresh}{\code{WifToken$refresh()}}
\item \href{#method-format}{\code{WifToken$format()}}
\item \href{#method-print}{\code{WifToken$print()}}
\item \href{#method-can_refresh}{\code{WifToken$can_refresh()}}
\item \href{#method-cache}{\code{WifToken$cache()}}
\item \href{#method-load_from_cache}{\code{WifToken$load_from_cache()}}
\item \href{#method-validate}{\code{WifToken$validate()}}
\item \href{#method-revoke}{\code{WifToken$revoke()}}
\item \href{#method-clone}{\code{WifToken$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="hash">}\href{../../httr/html/Token.html#method-hash}{\code{httr::Token$hash()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token2.0" data-id="sign">}\href{../../httr/html/Token2.0.html#method-sign}{\code{httr::Token2.0$sign()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Get a token via workload identity federation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$new(params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of parameters for \code{init_oauth_external_account()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A WifToken.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-init_credentials}{}}}
\subsection{Method \code{init_credentials()}}{
Enact the actual token exchange for workload identity
federation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$init_credentials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refreshes the token, which means re-doing the entire token
flow in this case.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format a \code{\link[=WifToken]{WifToken()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{\link[=WifToken]{WifToken()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-can_refresh"></a>}}
\if{latex}{\out{\hypertarget{method-can_refresh}{}}}
\subsection{Method \code{can_refresh()}}{
Placeholder implementation of required method. Returns \code{TRUE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$can_refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cache"></a>}}
\if{latex}{\out{\hypertarget{method-cache}{}}}
\subsection{Method \code{cache()}}{
Placeholder implementation of required method. Returns self.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_from_cache"></a>}}
\if{latex}{\out{\hypertarget{method-load_from_cache}{}}}
\subsection{Method \code{load_from_cache()}}{
Placeholder implementation of required method. Returns self.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$load_from_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Placeholder implementation of required method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-revoke"></a>}}
\if{latex}{\out{\hypertarget{method-revoke}{}}}
\subsection{Method \code{revoke()}}{
Placeholder implementation of required method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$revoke()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WifToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
